+++ lib/rubocop/cop/lint/unused_block_argument.rb
@@ -35,11 +35,11 @@ module RuboCop
           scope = variable.scope
           all_arguments = scope.variables.each_value.select(&:block_argument?)
 
-          if lambda?(scope.node)
-            message << message_for_lambda(variable, all_arguments)
-          else
-            message << message_for_normal_block(variable, all_arguments)
-          end
+          message << if lambda?(scope.node)
+                       message_for_lambda(variable, all_arguments)
+                     else
+                       message_for_normal_block(variable, all_arguments)
+                     end
 
           message
         end